Implementada a secao de investimentos para clientes e admins + correcao de alguns bugs de movimentacao

+++ serverIfBank/src/main/java/com/ifsp/ifBank/serverIfBank/repository/ContaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ContaRepository extends JpaRepository<Conta, Integer> {
-    Optional<Conta> findByUsuarioId(Long usuarioId);
+    Optional<Conta> findByUsuarioId(Integer usuarioId);
 }
